requirements.txt
matplotlib==3.8.0
obs-websocket-py==1.0
pytchat==0.5.5
requests==2.28.1
urllib3==1.26.10
websocket-client==1.7.0
websockets==11.0.3
pyyaml==6.0.2